requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
browser-cookie3==0.20.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dotenv
exceptiongroup==1.2.2
fastapi==0.115.6
flask==3.1.0
frozenlist==1.5.0
genshin==1.7.3
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
lz4==4.3.3
markupsafe==3.0.2
multidict==6.1.0
numpy==2.0.2
openperplex==0.2.2
pandas==2.2.3
pip
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic-core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tk
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
werkzeug==3.1.3
wheel==0.44.0
yarl==1.18.3
zipp==3.21.0
